refactor: extract sequence domain modules into sequence/ package
Move sequence_classes.py → sequence/classes.py and
participant.py → sequence/participant.py. Fix internal imports
and correct find_closest_participant to satisfy mypy.
