+++ src/plantuml_gui/sequence/classes.py
